requirements.txt
asgiref==3.2.7
Django==3.0.5
django-crispy-forms==1.10.0
django-filter==2.4.0
natsort==7.1.1
Pillow==8.2.0
pytz==2019.3
sqlparse==0.3.1
